CrewAI/MyCrewProject/Financial Researcher Project/requirements.txt
pydantic>=2.7,<2.9
pydantic-core>=2.18,<2.20
crewai
python-dotenv
crewai-tools



